adds climb card component, add more details to climb route show page, adds user climbs to route show page
